Bug 1653241 - Turn all the AOSP-specific libraries used by B2G into build dependencies r=glandium
This patch removes the configure flags used for adding AOSP-specific shared
objects and libraries and replaces them with a moz.build template. This allows
us to put the libraries and shared objects under a single folder in the
sysroot and makes it easier to add more. It also greatly simplifies the
mozconfig files.

To this end the script used to build the sysroot has been altered so that it
can be used for in-tree builds. The creation of the sysroot and its packaging
have been split and the result now preserves the various files' timestamps, to
minimize rebuilds when the sysroot is regenerated with the same files.

Differential Revision: https://phabricator.services.mozilla.com/D90683