Bug 1653198 - Implement bootstraping Boot2Gecko r=glandium
The current implementation is limited to Linux as we don't support building on
macOS and Windows for the time being. This patch also simplifies the default
configuration by removing unnecessary CPPFLAGS and LDFLAGS additions, and
getting rid of the --with-target-arch-name configure option which was needed
when we didn't have a proper sysroot.

Differential Revision: https://phabricator.services.mozilla.com/D96722