Bug 138941 - Dispatch visibilitychange event when isUnderHiddenEmbedderElement is also ready. r=andy.wu
App fail to focus an element even when its document is in visible state.

When focusing an element, gecko not only check its visibility, but also check
whether it's under a hidden embedder element. Upstream has exposed
`isUnderHiddenEmbedderElement` to BrowsingContext in MozBug#1776339, leading
the check of `nsIFrame::IsVisibleConsideringAncestors` looks for this value.

In BrowsingContext, state of `active`(will attempt dispatching visibilitychange
event) and state of `isUnderHiddenEmbedderElement` are not synced, so the fix
is to dispatch visibilitychange event when both value are set.